Add STTError type and improve recognition diagnostics
